requirement.txt
Package            Version
------------------ ---------
annotated-types    0.7.0
anyio              4.9.0
certifi            2025.4.26
charset-normalizer 3.4.2
distro             1.9.0
exceptiongroup     1.3.0
h11                0.16.0
httpcore           1.0.9
httpx              0.28.1
idna               3.10
jiter              0.10.0
openai             1.82.0
pip                22.0.4
pydantic           2.11.5
pydantic_core      2.33.2
python-dotenv      1.1.0
regex              2024.11.6
requests           2.32.3
setuptools         58.1.0
sniffio            1.3.1
tiktoken           0.9.0
tqdm               4.67.1
typing_extensions  4.13.2
typing-inspection  0.4.1
urllib3            2.4.0
